Crate pallet_aura_style_filter

A Nimbus filter for the AuRa consensus algorithm. This filter does not use any entropy, it simply rotates authors in order. A single author is eligible at each slot.

In the Substrate ecosystem, this algorithm is typically known as AuRa (authority round). There is a well known implementation in the main Substrate repository and published at https://crates.io/crates/sc-consensus-aura. There are two primary differences between the approaches:

This filter leverages all the heavy lifting of the Nimbus framework and consequently is capable of expressing Aura in < 100 lines of code.

Whereas sc-consensus-aura includes the entire consensus stack including block signing, digest formats, and slot prediction. This is a lot of overhead for a sipmle round robin consensus that basically boils down to this function https://github.com/paritytech/substrate/blob/0f849efc/client/consensus/aura/src/lib.rs#L91-L106

The Nimbus framework places the author checking logic in the runtime which makes it relatively easy for relay chain validators to confirm the author is valid.

Whereas sc-consensus-aura places the author checking offchain. The offchain approach is fine for standalone layer 1 blockchains, but net well suited for verification on the relay chain where validators only run a wasm blob.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.